Update AMR suite scripts

+++ orca-scripts/run_amr.sh
@@ -1,7 +1,13 @@
 #!/usr/bin/env bash
+
 set -eu
 OR_PREFIX=/users/ankushj/repos/orca-workspace/orca-install
-OR_UMB_PREFIX=/users/ankushj/repos/orca-workspace/orca-umb-install
+
+# for TAU
+MPI_HOME=/users/ankushj/amr-workspace/mvapich-install-ub22
+TAU_ROOT=/users/ankushj/repos/orca-workspace/tau-prefix/tau-2.34.1
+export PATH=$MPI_HOME/bin:$PATH
+export PATH=$PATH:${TAU_ROOT}/x86_64/bin
 
 # for do_mpirun and gen_hosts
 source $OR_PREFIX/scripts/common.sh
@@ -10,14 +16,16 @@ source $OR_PREFIX/scripts/orca_common.sh
 # Can use check_hosts.py to generate a hostfile on emulab
 # e.g. `check_hosts.py -o /tmp/hostfile.txt -e mon8`
 HOSTFILE=/tmp/hostfile.txt
-OR_SUITEDIR=/mnt/ltio/orcajobs/suites/amr-r128-psmdef
-OR_SUITE_DESC="
-AMR test suite with 128 ranks per node, default PSM settings.
 
-Goal is to see the performance impact of different AMR tunes
-on different ORCA modes.
-"
+# define outside:
+# OR_SUITEDIR: where to create per-run dirs
+# OR_SUITE_DESC: description of the suite
+# OR_ALL_ORCA_ENV, OR_ALL_MPI_ENV: ORCA and MPI environment variables
 
+# setup_suite_common: mostly just prevents accidental overwrites
+# - checks for -f flag to force overwrite of an existing suite dir
+# - do not set env vars here, as they will get reset,
+# set them in `setup_amr_common` instead
 setup_suite_common() {
   # die if suite dir already exists unless -f is passed
   local -i force=0
@@ -30,7 +38,16 @@ setup_suite_common() {
 
   shift $((OPTIND - 1))
 
-  rmdir $OR_SUITEDIR && true # try deleting if empty
+  # Add YYYYMMDD prefix to suite dir
+  local date_str=$(date +%Y%m%d)
+  local suite_basename=$(basename $OR_SUITEDIR)
+  local suite_dirname=$(dirname $OR_SUITEDIR)
+  local suite_namewdate=${date_str}_${suite_basename}
+  OR_SUITEDIR=${suite_dirname}/${suite_namewdate}
+
+  if [ -d $OR_SUITEDIR ]; then
+    rmdir $OR_SUITEDIR && true # try deleting if empty
+  fi
 
   if [ $force -eq 0 ] && [ -d $OR_SUITEDIR ]; then
     die "Suite directory already exists: $OR_SUITEDIR"
@@ -63,6 +80,24 @@ setup_gen_amrdeck() {
     -e "s/{nlim}/$nlim/g" >$OR_AMR_DECK
 }
 
+# setup_amr_common_add_env_vars: add all env vars to the environment
+# - used internally in setup_amr_common
+# - uses: OR_ALL_COMMON_ENV_VARS, OR_ALL_ORCA_ENV_VARS, OR_ALL_MPI_ENV_VARS
+setup_amr_common_add_env_vars() {
+  # add all common env vars
+  for key in "${!OR_ALL_ORCA_ENV[@]}"; do
+    local val=${OR_ALL_ORCA_ENV[$key]}
+    message "-INFO- Adding ORCA env var: $key = $val"
+    add_orca_env_var "$key" "$val"
+  done
+
+  for key in "${!OR_ALL_MPI_ENV[@]}"; do
+    local val=${OR_ALL_MPI_ENV[$key]}
+    message "-INFO- Adding MPI env var: $key = $val"
+    add_mpi_env_var "$key" "$val"
+  done
+}
+
 # setup_amr_common: common setup for AMR
 # - creates jobdir, prepares deck
 setup_amr_common() {
@@ -77,8 +112,6 @@ setup_amr_common() {
   # Bind AGG to specific CPU/MEM
   OR_AGG_BIN="numactl --physcpubind=0-7 --membind=0 $OR_AGG_BIN"
 
-  OR_AMR_BIN=$OR_UMB_PREFIX/bin/phoebus
-  OR_AMR_DECK_IN=$OR_UMB_PREFIX/decks/blast_wave_3d.512.pin.in
   # setup_gen_amrdeck <policy_name> <nlim>
   # policies: baseline, cdpc512par8, hybridX, lpt
   setup_gen_amrdeck $OR_AMR_POLICY $OR_AMR_NSTEPS
@@ -86,10 +119,10 @@ setup_amr_common() {
 
   OR_CFG_YAML="$OR_PREFIX/config/wfopts.yml" # ORCA config YAML
   OR_ORCA_ENABLED=1                          # enable ORCA by default
-}
 
-setup_profile_deftimeout() {
-  echo # does nothing
+  # add_common_env_var PSM_ERRCHK_TIMEOUT "1:4:1"
+  # add all common env vars
+  setup_amr_common_add_env_vars
 }
 
 # shorttimeout: set short timeout for PSM
@@ -114,42 +147,58 @@ set_new_yaml_with_cmdseq() {
 # noorca: disable ORCA
 setup_profile_noorca() {
   OR_ORCA_ENABLED=0
+  OR_RUN_TYPE="orca" # dir layout is still orc
 }
 
 # orcadisabled: load ORCA but disable tracers
 setup_profile_tracers_disabled() {
   local cmdseq="set-flow disable-tracers; resume"
   set_new_yaml_with_cmdseq "$cmdseq"
+  OR_RUN_TYPE="orca"
 }
 
 # tracendrop: trace and drop at MPI
 setup_profile_tracendrop() {
   local cmdseq="set-flow trace-and-drop; resume"
   set_new_yaml_with_cmdseq "$cmdseq"
+  OR_RUN_TYPE="orca"
 }
 
 # trace_mpip2p: trace mpi p2p messages
 setup_profile_trace_mpip2p() {
   local cmdseq="set-flow enable-tracers mpi_messages; resume"
   set_new_yaml_with_cmdseq "$cmdseq"
+  OR_RUN_TYPE="orca"
 }
 
 # trace_mpip2p_notest: trace mpi p2p messages but disable MPI_Test probe
 setup_profile_trace_mpip2p_notest() {
   local cmdseq="set-flow enable-tracers mpi_messags; disable-probe mpi_messages MPI_Test; resume"
   set_new_yaml_with_cmdseq "$cmdseq"
+  OR_RUN_TYPE="orca"
 }
 
-# trace_all: trace everything
+# trace_all: trace all tracers
 setup_profile_trace_all() {
   local cmdseq="set-flow enable-tracers; resume"
   set_new_yaml_with_cmdseq "$cmdseq"
+  OR_RUN_TYPE="orca"
+}
+
+# tau: run with TAU tracing=on, throttling=off
+setup_profile_tau() {
+  OR_ORCA_ENABLED=0
+  add_common_env_var TAU_TRACE 1
+  add_common_env_var TAU_THROTTLE 0
+  OR_RUN_TYPE="tau"
+  OR_MPI_BIN="tau_exec $OR_MPI_BIN"
 }
 
 # setup_profile: run ``
 setup_profile() {
   local profile=$1
-  local setup_func="setup_profile_$profile"
+  local profile_name=${profile#[0-9]*_} # strip leading digits and underscore
+  local setup_func="setup_profile_$profile_name"
 
   if declare -f "$setup_func" >/dev/null; then
     $setup_func
@@ -173,38 +222,27 @@ run_profile() {
 
   setup_amr_common       # common setup
   setup_profile $profile # profile-specific setup/overrides
-  run_orcaexp
+  run_main
   reset_all_env_vars # reset everything for next run
 }
 
-# extract_runtime: extract runtime from mpi.log from AMR
-# - generates: amr_runtimes.csv
-extract_runtime() {
-  local profile=$1
-  local amr_log=$OR_SUITEDIR/$profile/mpi.log
-  local runtime=$(cat $amr_log | grep "walltime used" | cut -d= -f2)
-  local runtime_fmt=$(echo $runtime | awk '{printf "%f\n", $1}')
-
-  echo "$profile runtime: $runtime_fmt seconds"
-  echo "$profile,$runtime_fmt" >>$OR_SUITEDIR/amr_runtimes.csv
-}
-
+# main: main function to run hardcoded profiles
 main() {
-  # CTL node is implied, and will be added if ORCA is enabled
-  OR_AGGCNT=1              # AGG count
-  OR_MPI_NNODES=8          # MPI nodes
-  OR_MPI_NRANKS=128        # MPI ranks per node
-  OR_FLOW_YAML=""          # Flow YAML is optional
-  OR_AMR_POLICY="baseline" # policy name
-  OR_AMR_NSTEPS=500        # timesteps
-
   profiles=("noorca" "tracers_disabled" "tracendrop"
-    "trace_mpip2p" "trace_mpip2p_notest" "trace_all")
+    "trace_mpip2p" "trace_mpip2p_notest" "trace_all" "tau")
   # profiles=("noorca" "tracers_disabled" "trace_all")
+  # profiles=("trace_mpip2p")
+  # profiles=("tau")
+  pidx=0 # profile_idx to prefix for sorted order
+
+  # Pick up env vars from outer script
+  source <(printf "%s" "$OR_ALL_ORCA_ENV_EXP")
+  source <(printf "%s" "$OR_ALL_MPI_ENV_EXP")
 
   for profile in "${profiles[@]}"; do
-    run_profile $profile
-    extract_runtime $profile
+    profile_name=${pidx}_${profile}
+    run_profile $profile_name
+    pidx=$((pidx + 1))
   done
 }
